Enum sqlite::Value

pub enum Value {
    Binary(Vec<u8>),
    Float(f64),
    Integer(i64),
    String(String),
    Null,
}

A typed value.

Variants
Binary(Vec<u8>)

Binary data.

Float(f64)

A floating-point number.

Integer(i64)

An integer number.

String(String)

A string.

Null

A null value.

Methods

impl Value

fn as_binary(&self) -> Option<&[u8]>

Return the binary data if the value is Binary.

fn as_float(&self) -> Option<f64>

Return the floating-point number if the value is Float.

fn as_integer(&self) -> Option<i64>

Return the integer number if the value is Integer.

fn as_string(&self) -> Option<&str>

Return the string if the value is String.

fn kind(&self) -> Type

Return the type.