Struct sqlite::Connection

pub struct Connection {
    // some fields omitted
}

A connection to a database.

Methods

impl Connection

fn open<T: AsRef<Path>>(path: T) -> Result<Connection>

Open a connection to a new or existing database.

fn execute<T: AsRef<str>>(&self, query: T) -> Result<()>

Execute a query without processing the resulting rows if any.

fn process<T: AsRef<str>, F>(&self, query: T, callback: F) -> Result<()> where F: FnMut(&[(&str, Option<&str>)]) -> bool

Execute a query and process the resulting rows if any.

The callback is triggered for each row. If the callback returns false, no more rows will be processed. For large queries and non-string data types, prepared statement are highly preferable; see prepare.

fn prepare<'l, T: AsRef<str>>(&'l self, query: T) -> Result<Statement<'l>>

Create a prepared statement.

fn set_busy_handler<F>(&mut self, callback: F) -> Result<()> where F: FnMut(usize) -> bool + 'static

Set a callback for handling busy events.

The callback is triggered when the database cannot perform an operation due to processing of some other request. If the callback returns true, the operation will be repeated.

fn set_busy_timeout(&mut self, milliseconds: usize) -> Result<()>

Set an implicit callback for handling busy events that tries to repeat rejected operations until a timeout expires.

fn remove_busy_handler(&mut self) -> Result<()>

Remove the callback handling busy events.