Struct sqlite::Statement

pub struct Statement<'l> {
    // some fields omitted
}

A prepared statement.

Methods

impl<'l> Statement<'l>

fn columns(&self) -> usize

Return the number of columns.

fn kind(&self, i: usize) -> Type

Return the type of a column.

The type is revealed after the first step has been taken.

fn bind<T: Parameter>(&mut self, i: usize, parameter: T) -> Result<()>

Bind the parameter at a specific location.

The leftmost location has the index 1.

fn read<T: Value>(&self, i: usize) -> Result<T>

Read the value stored in a specific column.

The leftmost column has the index 0.

fn step(&mut self) -> Result<State>

Evaluate the statement one step at a time.

The function should be called multiple times until State::Done is reached in order to evaluate the statement entirely.

fn reset(&mut self) -> Result<()>

Reset the statement.