Trait sqlite::BindableWithIndex

pub trait BindableWithIndex {
    // Required method
    fn bind<T: ParameterIndex>(self, _: &mut Statement<'_>, _: T) -> Result<()>;
}

A type suitable for binding to a prepared statement given a parameter index.

Required Methods

fn bind<T: ParameterIndex>(self, _: &mut Statement<'_>, _: T) -> Result<()>

Bind to a parameter.

In case of integer indices, the first parameter has index 1.