Crate sqlite

Interface to SQLite.

Example

use std::path::Path;

let database = sqlite::open(&Path::new(":memory:")).unwrap();

database.execute(r#"
    CREATE TABLE `users` (id INTEGER, name VARCHAR(255));
    INSERT INTO `users` (id, name) VALUES (1, 'Alice');
"#).unwrap();

database.process("SELECT * FROM `users`;", |pairs| {
    for &(column, value) in pairs.iter() {
        println!("{} = {}", column, value.unwrap());
    }
    true
}).unwrap();

Structs

Database
A connection to a database.

Error
An error.

Statement
A prepared statement.

Enums

Binding
A binding of a parameter of a prepared statement.

ErrorKind
An error kind.

State
A state of a prepared statement.

Traits

Value
A value stored in a result row of a query.

Functions

open
Open a connection to a new or existing database.

Type Definitions

Result
A result.