Struct sqlite::Connection

pub struct Connection { /* fields omitted */ }

A database connection.

Methods

impl Connection

fn open<T: AsRef<Path>>(path: T) -> Result<Connection>

Open a connection to a new or existing database.

fn execute<T: AsRef<str>>(&self, statement: T) -> Result<()>

Execute a statement without processing the resulting rows if any.

fn iterate<T: AsRef<str>, F>(&self, statement: T, callback: F) -> Result<()> where
F: FnMut(&[(&str, Option<&str>)]) -> bool,

Execute a statement and process the resulting rows as plain text.

The callback is triggered for each row. If the callback returns false, no more rows will be processed. For large queries and non-string data types, prepared statement are highly preferable; see prepare.

fn prepare<'l, T: AsRef<str>>(&'l self, statement: T) -> Result<Statement<'l>>

Create a prepared statement.

fn set_busy_handler<F>(&mut self, callback: F) -> Result<()> where
F: FnMut(usize) -> bool + Send + 'static,

Set a callback for handling busy events.

The callback is triggered when the database cannot perform an operation due to processing of some other request. If the callback returns true, the operation will be repeated.

fn set_busy_timeout(&mut self, milliseconds: usize) -> Result<()>

Set an implicit callback for handling busy events that tries to repeat rejected operations until a timeout expires.

fn remove_busy_handler(&mut self) -> Result<()>

Remove the callback handling busy events.