Struct sqlite::Statement

pub struct Statement<'l> {
    // some fields omitted
}

A prepared statement.

Methods

impl<'l> Statement<'l>

fn bind<T: Bindable>(&mut self, i: usize, parameter: T) -> Result<()>

Bind a value to a parameter.

The leftmost parameter has the index 1.

fn next(&mut self) -> Result<State>

Advance to the next state.

The function should be called multiple times until State::Done is reached in order to evaluate the statement entirely.

fn read<T: Readable>(&self, i: usize) -> Result<T>

Read a value from a column.

The leftmost column has the index 0.

fn reset(&mut self) -> Result<()>

Reset the statement.

fn columns(&self) -> usize

Return the number of columns.

fn kind(&self, i: usize) -> Type

Return the type of a column.

The type is revealed after the first step has been taken.

fn cursor(self) -> Cursor<'l>

Upgrade to a cursor.